dialog no longer bundled everywhere, hyperlink to implementations in typings comments
