+++ ui/analyse/src/study/relay/relayTourView.ts
@@ -10,7 +10,6 @@ import { defined, scrollToInnerSelector } from 'common';
 import StudyCtrl from '../studyCtrl';
 import { toggle } from 'common/controls';
 import * as xhr from 'common/xhr';
-import { domDialog } from 'common/dialog';
 
 export default function (ctrl: AnalyseCtrl): VNode | undefined {
   const study = ctrl.study;
@@ -154,7 +153,7 @@ const header = (relay: RelayCtrl, ctrl: AnalyseCtrl) => {
             attrs: dataIcon(licon.InfoCircle),
             hook: onInsert(el => {
               el.addEventListener('click', () => {
-                domDialog({
+                lichess.dialog.dom({
                   htmlText: `<h2>Broadcast notifications</h2>
 <p>Subscribe to be notified when each round starts. Make sure that bell or push notifications are
 enabled for broadcasts in your <a href="/account/preferences/notification">notification settings</a>.</p>`,
