+++ ui/puzzle/src/keyboard.ts
@@ -1,6 +1,5 @@
 import * as control from './control';
 import PuzzleCtrl from './ctrl';
-import { snabDialog } from 'common/dialog';
 
 export default (ctrl: PuzzleCtrl) =>
   lichess.mousetrap
@@ -35,7 +34,7 @@ export default (ctrl: PuzzleCtrl) =>
     .bind('B', () => ctrl.blindfold(!ctrl.blindfold()));
 
 export const view = (ctrl: PuzzleCtrl) =>
-  snabDialog({
+  lichess.dialog.snab({
     class: 'help',
     htmlUrl: '/training/help',
     onClose: () => ctrl.keyboardHelp(false),
