+++ ui/round/src/socket.ts
@@ -1,6 +1,5 @@
 import * as game from 'game';
 import throttle from 'common/throttle';
-import { domDialog } from 'common/dialog';
 import * as xhr from './xhr';
 import RoundController from './ctrl';
 import { defined } from 'common';
@@ -145,7 +144,7 @@ export function make(send: SocketSend, ctrl: RoundController): RoundSocket {
       }
     },
     simulEnd(simul: game.Simul) {
-      domDialog({
+      lichess.dialog.dom({
         htmlText:
           '<div><p>Simul complete!</p><br /><br />' +
           `<a class="button" href="/simul/${simul.id}">Back to ${simul.name} simul</a></div>`,
