+++ ui/site/src/forum.ts
@@ -1,26 +1,27 @@
 import * as xhr from 'common/xhr';
-import { domDialog } from 'common/dialog';
 
 lichess.load.then(() => {
   $('.forum')
     .on('click', 'a.delete', function (this: HTMLAnchorElement) {
       const link = this;
-      domDialog({
-        cash: $('.forum-delete-modal'),
-        attrs: { view: { action: link.href } },
-      }).then(dlg => {
-        $(dlg.view)
-          .find('form')
-          .attr('action', link.href)
-          .on('submit', function (this: HTMLFormElement, e: Event) {
-            e.preventDefault();
-            xhr.formToXhr(this);
-            $(link).closest('.forum-post').hide();
-            dlg.close();
-          });
-        $(dlg.view).find('form button.cancel').on('click', dlg.close);
-        dlg.showModal();
-      });
+      lichess.dialog
+        .dom({
+          cash: $('.forum-delete-modal'),
+          attrs: { view: { action: link.href } },
+        })
+        .then(dlg => {
+          $(dlg.view)
+            .find('form')
+            .attr('action', link.href)
+            .on('submit', function (this: HTMLFormElement, e: Event) {
+              e.preventDefault();
+              xhr.formToXhr(this);
+              $(link).closest('.forum-post').hide();
+              dlg.close();
+            });
+          $(dlg.view).find('form button.cancel').on('click', dlg.close);
+          dlg.showModal();
+        });
       return false;
     })
     .on('click', 'form.unsub button', function (this: HTMLButtonElement) {
