+++ ui/site/src/publicChats.ts
@@ -1,5 +1,4 @@
 import { text, form } from 'common/xhr';
-import { domDialog } from 'common/dialog';
 
 lichess.load.then(() => {
   let autoRefreshEnabled = true;
@@ -40,7 +39,7 @@ lichess.load.then(() => {
 
     $('#communication').on('click', '.line:not(.lichess)', function (this: HTMLDivElement) {
       const $l = $(this);
-      domDialog({ cash: $('.timeout-modal') }).then(dlg => {
+      lichess.dialog.dom({ cash: $('.timeout-modal') }).then(dlg => {
         $('.username', dlg.view).text($l.find('.user-link').text());
         $('.text', dlg.view).text($l.text().split(' ').slice(1).join(' '));
         $('.button', dlg.view).on('click', function (this: HTMLButtonElement) {
